(deploy reorg/refactor) Reorganization in tests
- deployment.*.lists tests
- move js specific stuff: test/unit/deploy => test/unit/deployment/javascript
