+++ test/.jshintrc
@@ -20,21 +20,31 @@
     "acorn": true,
     "async": true,
     "bindings": true,
+    "builtInRulesCount": true,
     "cli": true,
+    "codeContents": true,
     "commands": true,
     "controller": true,
     "copy": true,
     "cp": true,
+    "createTemporaryDeployCode": true,
     "Daemon": true,
+    "DEPLOY_FILE": true,
+    "DEPLOY_DIR": true,
+    "deleteTemporaryDeployCode": true,
     "deploy": true,
-    "deployment": true,
+    "deployFile": true,
+    "deployFolder": true,
     "deployLists": true,
+    "deployment": true,
+    "deployTestCode": true,
     "discover": true,
     "document": true,
     "Duplex": true,
     "Emitter": true,
     "events": true,
     "exports": true,
+    "extract": true,
     "fs": true,
     "fsTemp": true,
     "Ignore": true,
@@ -46,7 +56,6 @@
     "mdns": true,
     "mkdirp": true,
     "NodeRSA": true,
-    "RSA": true,
     "osenv": true,
     "os": true,
     "path": true,
@@ -54,9 +63,12 @@
     "Project": true,
     "Promise": true,
     "provision": true,
+    "reference": true,
     "RemoteProcessSimulator": true,
     "request": true,
     "Request": true,
+    "RSA": true,
+    "sandbox": true,
     "sinon": true,
     "ssh": true,
     "sshpk": true,
