+++ test/common/bootstrap.js
@@ -74,3 +74,103 @@ global.USB = usb.USB;
 
 global.Request = function Request() {};
 util.inherits(global.Request, global.Stream);
+
+
+
+// Deployment Utilities, shared across
+//
+// - test/unit/deploy.js
+// - test/unit/deployment/javascript.js
+// - test/unit/deployment/python.js
+// - test/unit/deployment/rust.js
+//
+
+global.DEPLOY_DIR = path.join(process.cwd(), 'test/unit/', 'tmp');
+global.DEPLOY_FILE = path.join(global.DEPLOY_DIR, 'app.js');
+global.codeContents = 'console.log("testing deploy");';
+global.reference = new Buffer(global.codeContents);
+
+
+global.deployTestCode = function(tessel, opts, callback) {
+  // Create the temporary folder with example code
+  createTemporaryDeployCode()
+    .then(() => {
+
+      function closeAdvance(event) {
+        if (event === 'close') {
+          setImmediate(() => {
+            // Emit the close event to keep it going
+            tessel._rps.emit('close');
+          });
+        }
+      }
+
+      // When we get a listener that the Tessel process needs to close before advancing
+      tessel._rps.on('newListener', closeAdvance);
+
+      // Actually deploy the script
+      tessel.deploy({
+          entryPoint: path.relative(process.cwd(), global.DEPLOY_FILE),
+          push: opts.push,
+          single: opts.single
+        })
+        // If it finishes, it was successful
+        .then(() => {
+          tessel._rps.removeListener('newListener', closeAdvance);
+          callback();
+        })
+        // If not, there was an issue
+        .catch(callback);
+    });
+};
+
+global.createTemporaryDeployCode = function() {
+  return new Promise((resolve, reject) => {
+    mkdirp(global.DEPLOY_DIR, (err) => {
+      if (err) {
+        return reject(err);
+      } else {
+        fs.writeFile(global.DEPLOY_FILE, codeContents, (err) => {
+          if (err) {
+            reject(err);
+          } else {
+            resolve();
+          }
+        });
+      }
+    });
+  });
+};
+
+global.deleteTemporaryDeployCode = function() {
+  return new Promise(function(resolve, reject) {
+    fs.remove(global.DEPLOY_DIR, (err) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve();
+      }
+    });
+  });
+};
+
+global.extract = function(bundle, callback) {
+  var parser = tar.Parse();
+  var entries = [];
+
+  parser.on('entry', (entry) => {
+    if (entry.type === 'File') {
+      entries.push(entry.path);
+    }
+  });
+
+  parser.on('end', () => {
+    callback(null, entries);
+  });
+
+  parser.on('error', (error) => {
+    callback(error, null);
+  });
+
+  parser.end(bundle);
+};
